+++ src/base/CoincidenceGrouper.cpp
@@ -17,7 +17,7 @@ void CoincidenceGrouper::report()
 {
 	printf(">> CoincidenceGrouper report\n");
 	printf(" prompts passed\n");
-	printf("  %10u \n", nPrompts);
+	printf("  %10lu \n", nPrompts);
 	UnorderedEventHandler<GammaPhoton, Coincidence>::report();
 }
 
@@ -28,7 +28,7 @@ EventBuffer<Coincidence> * CoincidenceGrouper::handleEvents(EventBuffer<GammaPho
 	unsigned N =  inBuffer->getSize();
 	EventBuffer<Coincidence> * outBuffer = new EventBuffer<Coincidence>(N, inBuffer);
 
-	u_int32_t lPrompts = 0;
+	u_int64_t lPrompts = 0;
 	for(unsigned i = 0; i < N; i++) {
 		GammaPhoton &photon1 = inBuffer->get(i);
 		for(unsigned j = i+1; j < N; j++) {
